+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend_enclave_functions.go
@@ -7,6 +7,8 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/object_attributes_provider/label_key_consts"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/object_attributes_provider/label_value_consts"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/partition"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"strings"
@@ -293,6 +295,28 @@ func (backend *DockerKurtosisBackend) DestroyEnclaves(
 	return successfulEnclaveIds, erroredEnclaveIds, nil
 }
 
+// Register file artifacts using the provided URLs
+func (backend *DockerKurtosisBackend) RegisterFileArtifacts(
+	ctx context.Context,
+	fileArtifactsUrls map[service.FilesArtifactID]string,
+)(
+	resultErr error,
+) {
+	panic("Implement me")
+}
+
+// Repartition the enclave network
+func (backendCore *DockerKurtosisBackend) CreateRepartition(
+	ctx context.Context,
+	partitions []*partition.Partition,
+	newPartitionConnections map[partition.PartitionConnectionID]partition.PartitionConnection,
+	newDefaultConnection partition.PartitionConnection,
+)(
+	resultErr error,
+) {
+	panic("Implement me")
+}
+
 // ====================================================================================================
 // 									   Private helper methods
 // ====================================================================================================
@@ -388,3 +412,4 @@ func (backend *DockerKurtosisBackend) getEnclaveContainers(
 	}
 	return containers, nil
 }
+