+++ container-engine-lib/lib/backend_interface/kurtosis_backend.go
@@ -5,7 +5,12 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/api_container"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/engine"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/module"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/partition"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
+	"io"
+	"net"
 )
 
 // KurtosisBackend abstracts a Kurtosis backend, which will be a container engine (Docker or Kubernetes).
@@ -89,6 +94,24 @@ type KurtosisBackend interface {
 		resultErr error,
 	)
 
+	// Register some file artifacts which can be used for the enclave's services and modules
+	RegisterFileArtifacts(
+		ctx context.Context,
+		fileArtifactsUrls map[service.FilesArtifactID]string,
+	)(
+		resultErr error,
+	)
+
+	// Repartition the Enclave network defining which services will be on each part
+	CreateRepartition(
+		ctx context.Context,
+		partitions []*partition.Partition,
+		newPartitionConnections map[partition.PartitionConnectionID]partition.PartitionConnection,
+		newDefaultConnection partition.PartitionConnection,
+	)(
+		resultErr error,
+	)
+
 	CreateAPIContainer(
 		ctx context.Context,
 		image string,
@@ -107,7 +130,7 @@ type KurtosisBackend interface {
 		ctx context.Context,
 		filters *api_container.APIContainerFilters,
 	) (
-		// Matching API containers, keyed by their enclave ID
+	// Matching API containers, keyed by their enclave ID
 		map[string]*api_container.APIContainer,
 		error,
 	)
@@ -117,7 +140,7 @@ type KurtosisBackend interface {
 		ctx context.Context,
 		filters *enclave.EnclaveFilters,
 	) (
-		// Successful & errored API containers are keyed by their enclave ID
+	// Successful & errored API containers are keyed by their enclave ID
 		successApiContainerIds map[string]bool,
 		erroredApiContainerIds map[string]error,
 		resultErr error,
@@ -128,9 +151,114 @@ type KurtosisBackend interface {
 		ctx context.Context,
 		filters *enclave.EnclaveFilters,
 	) (
-		// Successful & errored API containers are keyed by their enclave ID
+	// Successful & errored API containers are keyed by their enclave ID
 		successApiContainerIds map[string]bool,
 		erroredApiContainerIds map[string]error,
+		resultErr error,
+	)
+
+	// Create a module from a container image with serialized params
+	CreateModule(
+		ctx context.Context,
+		id string,
+		containerImageName string,
+		serializedParams string,
+	)(
+		privateIp net.IP,
+		privatePort *port_spec.PortSpec,
+		publicIp net.IP,
+		publicPort *port_spec.PortSpec,
+		resultErr error,
+	)
+
+	// Gets modules using the given filters, returning a map of matched modules identified by their module ID
+	GetModules(ctx context.Context, filters *module.ModuleFilters) (map[string]*module.Module, error)
+
+	// Destroys the modules with the given filters, regardless of if they're running or not
+	DestroyModules(
+		ctx context.Context,
+		filters *module.ModuleFilters,
+	) (
+		successfulModuleIds map[string]bool, // "set" of module IDs that were successfully destroyed
+		erroredModuleIds map[string]error, // "set" of module IDs that errored when destroying, with the error
+		resultErr error, // Represents an error with the function itself, rather than the modules
+	)
+
+	// Creates a user service inside an enclave with the given configuration
+	CreateUserService(
+		ctx context.Context,
+		id string,
+		containerImageName string,
+		privatePorts []*port_spec.PortSpec,
+		entrypointArgs []string,
+		cmdArgs []string,
+		envVars map[string]string,
+		enclaveDataDirMntDirpath string,
+		filesArtifactMountDirpaths map[string]string,
+    )(
+		maybePublicIpAddr net.IP, // The ip exposed in the host machine. Will be nil if the service doesn't declare any private ports
+		publicPorts map[string]*port_spec.PortSpec, //Mapping of port-used-by-service -> port-on-the-host-machine where the user can make requests to the port to access the port. If a used port doesn't have a host port bound, then the value will be nil.
+		resultErr error,
+	)
+
+	// Gets user services using the given filters, returning a map of matched user services identified by their ID
+	GetUserServices(ctx context.Context, filters *service.ServiceFilters) (map[string]*service.Service, error)
+
+	// Get user service logs using the given filters, returning a map of matched user services identified by their ID and a readCloser object for each one
+	GetUserServiceLogs(ctx context.Context, filters *service.ServiceFilters) (map[string]io.ReadCloser, error)
+
+	// Executes a shell command inside an user service instance indenfified by its ID
+	RunUserServiceExecCommand (
+		ctx context.Context,
+		serviceId string,
+		commandArgs []string,
+	)(
+		exitCode int32,
+		output string,
+		resultErr error,
+	)
+
+	// Wait for succesful http endpoint response which can be used to check if the service is available
+	WaitForUserServiceHttpEndpointAvailability(
+		ctx context.Context,
+		serviceId string,
+		httpMethod string, //The httpMethod used to execute the request. Valid values: GET and POST
+		port uint32, //The port of the service to check. For instance 8080
+		path string, //The path of the service to check. It mustn't start with the first slash. For instance `service/health`
+		requestBody string, //The content of the request body. Only valid when the httpMethod is POST
+		bodyText string, //If the endpoint returns this value, the service will be marked as available (e.g. Hello World).
+		initialDelayMilliseconds uint32, //The number of milliseconds to wait until executing the first HTTP call
+		retries uint32, //Max number of HTTP call attempts that this will execute until giving up and returning an error
+		retriesDelayMilliseconds uint32, //Number of milliseconds to wait between retries
+	)(
+		resultErr error,
+	)
+
+	// Stop user services using the given filters,
+	StopUserServices(
+		ctx context.Context,
+		filters *service.ServiceFilters,
+	)(
+		successfulUserServiceIds map[string]bool, // "set" of user service IDs that were successfully stopped
+		erroredUserServiceIds map[string]error, // "set" of user service IDs that errored when stopping, with the error
+		resultErr error, // Represents an error with the function itself, rather than the user services
+	)
+
+	// Destroy user services using the given filters,
+	DestroyUserServices(
+		ctx context.Context,
+		filters *service.ServiceFilters,
+	)(
+		successfulUserServiceIds map[string]bool, // "set" of user service IDs that were successfully destroyed
+		erroredUserServiceIds map[string]error, // "set" of user service IDs that errored when destroying, with the error
+		resultErr error, // Represents an error with the function itself, rather than the user services
+	)
+
+	// Get an interactive shell to execute commands in an user service
+	GetShellOnUserService(
+		ctx context.Context,
+		userServiceId string,
+	)(
 		resultErr error,
 	)
 
@@ -146,19 +274,4 @@ type KurtosisBackend interface {
 
 	// TODO RunReplExecCommand
 
-	// TODO CreateModule
-
-	// TODO DestroyModule
-
-	// TODO CreateUserService
-
-	// TODO GetUserServices
-
-	// TODO GetUserServiceLogs
-
-	// TODO StopUserServices
-
-	// TODO GetShellOnUserService
-
-	// TODO DestroyUserServices
 }