+++ container-engine-lib/lib/backend_interface/objects/module/module.go
@@ -1,0 +1,15 @@
+package module
+
+// Object that represents POINT-IN-TIME information about a Kurtosis Module
+// Store this object and continue to reference it at your own risk!!!
+type Module struct {
+	id string
+}
+
+func NewModule(id string) *Module {
+	return &Module{id: id}
+}
+
+func (module *Module) GetID() string {
+	return module.id
+}