+++ container-engine-lib/docs/changelog.md
@@ -1,4 +1,7 @@
 # TBD
+### Features
+* Added `UserService` methods, `Modules` methods and `CreateRepartition` method in `KurtosisBackend` interface
+* Stubbing out methods for `UserService`, `Modules` and `CreateRepartition` into Docker implementation
 
 # 0.10.1
 ### Features